Corbi-package {Corbi}	R Documentation

Corbi - Bioinformatics analysis tools

Description

This pakcage provides a bundle of bioinformatics analysis tools for biomedical research.

Details

This is a collection of bioinformatics analysis tools developed by WuLab at Academy of Mathematics and Systems Science, Chinese Academy of Sciences.

Network comparison:

net_query Network querying method based on conditional random fields

net_query_batch Batch processing version of net_query

net_align Network alignment method based on conditional random fields

References

Qiang Huang, Ling-Yun Wu, and Xiang-Sun Zhang. An Efficient Network Querying Method Based on Conditional Random Fields. Bioinformatics, 27(22):3173-3178, 2011.

Qiang Huang, Ling-Yun Wu, and Xiang-Sun Zhang. Corbi: A new R package for biological network alignment and querying. BMC Systems Biology, 7(Suppl 2):S6, 2013.

[Package Corbi version 0.3-4 Index]