column {Corbi}	R Documentation

Extract a column from a matrix

Description

Extract a specified column from a sparse matrix rapidly

Usage

column(m, i)

Arguments

m
The matrix

i
The column index

Details

This function implements faster column extraction algorithm for the CsparseMatrix class in the package Matrix.

Value

This function will return the specified column as a vector of corresponding type.

[Package Corbi version 0.3-4 Index]