nnzero {Corbi}	R Documentation

The number of non-zero values of a submatrix

Description

Retuen the number of non-zero values of the specified submatrix of a given sparse matrix rapidly

Usage

nnzero(m, rows = 1:dim(m)[1], cols = 1:dim(m)[2])

Arguments

m
The matrix

rows
The integer vector of row index(es) or logical vector indicated the selected rows

cols
The integer vector of column index(es) or logical vector indicated the selected cols

Details

This function implements faster calculation algorithm for the CsparseMatrix and RsparseMatrix class in the package Matrix.

Value

This function will return the number of non-zero values in the specified submatrix.

[Package Corbi version 0.3-4 Index]