read_net {Corbi}	R Documentation

Read network information from text file

Description

Read the network information from a text file with specific format.

Usage

read_net(file)

Arguments

file
The name of text file

Details

This function reads the network information from a text file with specific format: each line contains two strings separated by spaces, which correspond to the names of two end points of one edge in the network.

Value

A list with the following components:

size
The number of network nodes

node
The vector of network node names

matrix
The logical adjacency matrix

See Also

write_net

[Package Corbi version 0.3-4 Index]