submatrix {Corbi}	R Documentation

Extract a submatrix from a matrix

Description

Extract a specified submatrix from a sparse matrix rapidly

Usage

submatrix(m, rows, cols)

Arguments

m
The matrix

rows
The integer vectors of row index(es)

cols
The integer vectors of column index(es)

Details

This function implements faster submatrix extraction algorithm for the CsparseMatrix class in the package Matrix.

Value

This function will return the specified submatrix as a matrix of corresponding type.

[Package Corbi version 0.3-4 Index]