write_net {Corbi}	R Documentation

Write network information to text file

Description

Write the network information to a text file with specific format.

Usage

write_net(net, file)

Arguments

net
A list as returned by read_net

file
The name of text file

Details

This function writes the network information to a text file with specific format: each line contains two strings separated by spaces, which correspond to the names of two end points of one edge in the network.

See Also

read_net

[Package Corbi version 0.3-4 Index]